refactor(vadRecorder): enhance startActiveListening to pass onSpeechEnd callback
